refactor: move isAnySegmentReachable to code-path-utils
