+++ lib/rules/array-callback-return.js
@@ -10,6 +10,7 @@
 //------------------------------------------------------------------------------
 
 const astUtils = require("./utils/ast-utils");
+const { isAnySegmentReachable } = require("./utils/code-path-utils");
 
 //------------------------------------------------------------------------------
 // Helpers
@@ -353,7 +354,7 @@ module.exports = {
 			} else {
 				if (
 					node.body.type === "BlockStatement" &&
-					astUtils.isAnySegmentReachable(funcInfo.currentSegments)
+					isAnySegmentReachable(funcInfo.currentSegments)
 				) {
 					messageAndSuggestions.messageId = funcInfo.hasReturn
 						? "expectedAtEnd"
