+++ lib/rules/getter-return.js
@@ -10,6 +10,7 @@
 //------------------------------------------------------------------------------
 
 const astUtils = require("./utils/ast-utils");
+const { isAnySegmentReachable } = require("./utils/code-path-utils");
 
 //------------------------------------------------------------------------------
 // Helpers
@@ -83,7 +84,7 @@ module.exports = {
 		function checkLastSegment(node) {
 			if (
 				funcInfo.shouldCheck &&
-				astUtils.isAnySegmentReachable(funcInfo.currentSegments)
+				isAnySegmentReachable(funcInfo.currentSegments)
 			) {
 				context.report({
 					node,
