+++ lib/rules/no-fallthrough.js
@@ -9,7 +9,7 @@
 //------------------------------------------------------------------------------
 
 const { directivesPattern } = require("../shared/directives");
-const astUtils = require("./utils/ast-utils");
+const { isAnySegmentReachable } = require("./utils/code-path-utils");
 
 //------------------------------------------------------------------------------
 // Helpers
@@ -225,7 +225,7 @@ module.exports = {
 				 * `break`, `return`, or `throw` are unreachable.
 				 * And allows empty cases and the last case.
 				 */
-				const isSwitchExitReachable = astUtils.isAnySegmentReachable(
+				const isSwitchExitReachable = isAnySegmentReachable(
 					currentCodePathSegments,
 				);
 				const isFallthrough =
