+++ lib/rules/no-unreachable-loop.js
@@ -5,7 +5,7 @@
 
 "use strict";
 
-const astUtils = require("./utils/ast-utils");
+const { isAnySegmentReachable } = require("./utils/code-path-utils");
 
 //------------------------------------------------------------------------------
 // Helpers
@@ -162,7 +162,7 @@ module.exports = {
 				 * Ignore unreachable loop statements to avoid unnecessary complexity in the implementation, or false positives otherwise.
 				 * For unreachable segments, the code path analysis does not raise events required for this implementation.
 				 */
-				if (astUtils.isAnySegmentReachable(currentCodePathSegments)) {
+				if (isAnySegmentReachable(currentCodePathSegments)) {
 					loopsToReport.add(node);
 				}
 			},
