+++ lib/rules/no-useless-return.js
@@ -10,6 +10,7 @@
 
 const astUtils = require("./utils/ast-utils"),
 	FixTracker = require("./utils/fix-tracker");
+const { isAnySegmentReachable } = require("./utils/code-path-utils");
 
 //------------------------------------------------------------------------------
 // Helpers
@@ -331,7 +332,7 @@ module.exports = {
 					astUtils.isInLoop(node) ||
 					isInFinally(node) ||
 					// Ignore `return` statements in unreachable places (https://github.com/eslint/eslint/issues/11647).
-					!astUtils.isAnySegmentReachable(scopeInfo.currentSegments)
+					!isAnySegmentReachable(scopeInfo.currentSegments)
 				) {
 					return;
 				}
