+++ lib/rules/utils/ast-utils.js
@@ -1436,21 +1436,6 @@ let needsPrecedingSemicolon;
 }
 
 /**
- * Checks all segments in a set and returns true if any are reachable.
- * @param {Set<CodePathSegment>} segments The segments to check.
- * @returns {boolean} True if any segment is reachable; false otherwise.
- */
-function isAnySegmentReachable(segments) {
-	for (const segment of segments) {
-		if (segment.reachable) {
-			return true;
-		}
-	}
-
-	return false;
-}
-
-/**
  * Checks if a node is used as an import attribute key, either in a static or dynamic import.
  * @param {ASTNode} node The node to check.
  * @returns {boolean} Whether the node is used as an import attribute key.
@@ -2891,6 +2876,5 @@ module.exports = {
 	canContinueExpressionInClassBody,
 	needsPrecedingSemicolon,
 	isImportAttributeKey,
-	isAnySegmentReachable,
 	getOpeningParenOfParams,
 };
