+++ tests/lib/rules/utils/ast-utils.js
@@ -2578,41 +2578,3 @@ describe("ast-utils", () => {
 		});
 	});
 });
-
-describe("isAnySegmentReachable", () => {
-	it("should return false for an empty set of segments", () => {
-		const segments = new Set();
-
-		assert.strictEqual(astUtils.isAnySegmentReachable(segments), false);
-	});
-
-	it("should return true when a single segment is reachable", () => {
-		const segments = new Set([{ reachable: true }]);
-
-		assert.strictEqual(astUtils.isAnySegmentReachable(segments), true);
-	});
-
-	it("should return false when a single segment is unreachable", () => {
-		const segments = new Set([{ reachable: false }]);
-
-		assert.strictEqual(astUtils.isAnySegmentReachable(segments), false);
-	});
-
-	it("should return false when all segments are unreachable", () => {
-		const segments = new Set([{ reachable: false }, { reachable: false }]);
-
-		assert.strictEqual(astUtils.isAnySegmentReachable(segments), false);
-	});
-
-	it("should return true when all segments are reachable", () => {
-		const segments = new Set([{ reachable: true }, { reachable: true }]);
-
-		assert.strictEqual(astUtils.isAnySegmentReachable(segments), true);
-	});
-
-	it("should return true when at least one segment is reachable", () => {
-		const segments = new Set([{ reachable: false }, { reachable: true }]);
-
-		assert.strictEqual(astUtils.isAnySegmentReachable(segments), true);
-	});
-});
